Make curios compat safer
