+++ src/main/java/net/silvertide/artifactory/Artifactory.java
@@ -5,15 +5,12 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.silvertide.artifactory.client.events.TooltipHandler;
-import net.silvertide.artifactory.compat.CuriosCompat;
 import net.silvertide.artifactory.config.Config;
 import net.silvertide.artifactory.registry.*;
 import net.silvertide.artifactory.setup.CuriosSetup;
@@ -22,7 +19,6 @@ import org.slf4j.Logger;
 // Future Features
 // TODO: Make a keybind that searches nearby players and pulls your items into your inventory.
 // TODO: Search through all containers if theyre in the inventory as well, like backpacks, for soulbound.
-// TODO: Add curios slots that only allow artifacts placed in them for storage.
 
 @Mod(Artifactory.MOD_ID)
 public class Artifactory
@@ -31,8 +27,6 @@ public class Artifactory
     public static final Logger LOGGER = LogUtils.getLogger();
     public Artifactory() {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-
-        modEventBus.addListener(this::commonSetup);
 
         AttributeRegistry.register(modEventBus);
         ItemRegistry.register(modEventBus);
@@ -45,7 +39,6 @@ public class Artifactory
 
         MinecraftForge.EVENT_BUS.register(this);
     }
-    private void commonSetup(final FMLCommonSetupEvent event) {}
 
     @Mod.EventBusSubscriber(modid = MOD_ID, bus=Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
     public static class ClientModEvents {
