+++ src/main/java/net/silvertide/artifactory/compat/CompatFlags.java
@@ -1,0 +1,7 @@
+package net.silvertide.artifactory.compat;
+
+public final class CompatFlags {
+    private CompatFlags() {}
+
+    public static boolean CURIOS_LOADED = false;
+}
