+++ src/main/java/net/silvertide/artifactory/util/AttunementService.java
@@ -1,8 +1,8 @@
 package net.silvertide.artifactory.util;
 
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.silvertide.artifactory.compat.CompatFlags;
 import net.silvertide.artifactory.config.Config;
 import net.silvertide.artifactory.storage.ArtifactorySavedData;
 import net.silvertide.artifactory.storage.AttunedItem;
@@ -10,7 +10,7 @@ import net.silvertide.artifactory.storage.AttunedItem;
 public final class AttunementService {
     private AttunementService() {}
 
-    // Atunement Actions
+    // Attunement Actions
     public static void increaseLevelOfAttunement(ServerPlayer player, ItemStack stack) {
         int levelOfAttunementAchieved = AttunementUtil.getLevelOfAttunementAchieved(stack);
         boolean successfulAttunementIncrease = false;
@@ -66,7 +66,10 @@ public final class AttunementService {
         for (ItemStack item : player.getInventory().offhand) {
             clearBrokenAttunementIfExists(item);
         }
-        AttunementUtil.ejectInvalidCurios(player);
+
+        if(CompatFlags.CURIOS_LOADED) {
+            net.silvertide.artifactory.compat.CuriosCompat.ejectInvalidCurios(player);
+        }
     }
 
     // The purpose of this method is to check if the itemstack is attuned to a player
